+++ packages/chakra-ui-core/src/CFormControl/CFormControl.stories.js
@@ -19,7 +19,7 @@ storiesOf('UI | FormControl', module)
           <c-input-group>
             <c-input-left-element color="gray.300" fontSize="1.2em"><c-icon name="lock" /></c-input-left-element>
             <c-input v-model="pw" :type="shouldShowPassword ? 'text' : 'password'" placeholder="Password" />
-            <c-input-right-element @click.native="shouldShowPassword = !shouldShowPassword" ><c-icon :name="shouldShowPassword ? 'eye-slash' : 'eye'" color="gray.500" /></c-input-right-element>
+            <c-input-right-element @click.native="shouldShowPassword = !shouldShowPassword" ><c-icon :name="shouldShowPassword ? 'eye-off' : 'eye'" color="gray.500" /></c-input-right-element>
           </c-input-group>
           <c-form-helper-text>Enter your password</c-form-helper-text>
         </c-form-control>
@@ -40,7 +40,7 @@ storiesOf('UI | FormControl', module)
           <CInputGroup>
             <CInputLeftElement color="gray.300" fontSize="1.2em"><CIcon name="lock" /></CInputLeftElement>
             <CInput v-model="value" :type="shouldShowPassword ? 'text' : 'password'" placeholder="Password" />
-            <CInputRightElement @click.native="shouldShowPassword = !shouldShowPassword" ><CIcon :name="shouldShowPassword ? 'eye-slash' : 'eye'" color="gray.500" /></CInputRightElement>
+            <CInputRightElement @click.native="shouldShowPassword = !shouldShowPassword" ><CIcon :name="shouldShowPassword ? 'eye-off' : 'eye'" color="gray.500" /></CInputRightElement>
           </CInputGroup>
           <CFormErrorMessage v-show="value.length < 4" id="url-error">Your password is too short.</CFormErrorMessage>
         </CFormControl>