+++ packages/chakra-ui-core/src/CPopper/utils/popper.styles.js
@@ -1,4 +1,4 @@
-import { css } from 'emotion'
+import { css } from '@emotion/css'
 
 /**
  * Evaluates popper arrow styles