+++ website/components/MobileNav.vue
@@ -5,7 +5,7 @@
       aria-label="Navigation Menu"
       font-size="20px"
       variant="ghost"
-      icon="bars"
+      icon="menu"
       @click="toggleMenu"
     />
     <CDrawer size="xs" :is-open="isOpen" placement="left" :on-close="handleClose">