+++ website/plugins/editor.js
@@ -1,8 +1,7 @@
 import Vue from 'vue'
 import VueLive from 'vue-live'
 import 'vue-prism-editor/dist/VuePrismEditor.css'
-import LiveEditor from '~/components/LiveEditor.js'
-import CodeBlock from '@/components/CodeBlock'
+import { CodeBlock, LiveEditor } from '~/components/code'
 
 Vue.use(VueLive)
 Vue.component(LiveEditor)